chore: llvm ir makefile

+++ runtime/src/io.zig
@@ -1,5 +1,5 @@
 const std = @import("std");
 
-pub fn naviary_print(value: c_int) void {
+pub fn naviary_print(value: i64) void {
     std.debug.print("{d}\n", .{value});
 }
