+++ runtime/src/lib.zig
@@ -1,6 +1,6 @@
 const io = @import("io.zig");
 
 // TODO: add types
-pub export fn print(value: c_int) void {
+pub export fn print(value: i64) void {
     io.naviary_print(value);
 }
